refactor(frontend): split admin console components
