+++ frontend/components/admin/admin-kpi-card.tsx
@@ -1,0 +1,44 @@
+import { HugeiconsIcon } from "@hugeicons/react"
+import type { IconSvgElement } from "@hugeicons/react"
+
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import { Skeleton } from "@/components/ui/skeleton"
+
+interface AdminKpiCardProps {
+  title: string
+  value: string | number
+  description: string
+  icon: IconSvgElement
+  isLoading: boolean
+}
+
+export function AdminKpiCard({
+  title,
+  value,
+  description,
+  icon,
+  isLoading,
+}: AdminKpiCardProps) {
+  return (
+    <Card>
+      <CardHeader className="flex flex-row items-center justify-between pb-2">
+        <CardTitle className="text-sm font-medium text-muted-foreground">
+          {title}
+        </CardTitle>
+        <HugeiconsIcon
+          icon={icon}
+          className="size-4 text-muted-foreground"
+          strokeWidth={1.5}
+        />
+      </CardHeader>
+      <CardContent>
+        {isLoading ? (
+          <Skeleton className="h-8 w-24" />
+        ) : (
+          <div className="text-2xl font-bold tracking-tight">{value}</div>
+        )}
+        <p className="mt-1 text-xs text-muted-foreground">{description}</p>
+      </CardContent>
+    </Card>
+  )
+}
